backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic==2.9.2
pydantic-settings==2.6.0
alembic==1.13.3
celery==5.4.0
redis==5.1.1
python-multipart==0.0.12
httpx==0.27.2
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml_html_clean==0.2.2
playwright==1.48.0
pymupdf==1.24.12
pytesseract==0.3.13
Pillow==11.0.0
minio==7.2.10
reportlab==4.2.5
prometheus-client==0.21.0
openai==1.54.4
python-dateutil==2.9.0.post0
pytest==8.3.3
